@@ -22,32 +22,28 @@
   },
   "private": true,
   "dependencies": {
-    "@emotion/react": "11.9.0",
-    "@emotion/styled": "11.8.1",
+    "@babel/preset-env": "^7.17.10",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
     "@eva-design/eva": "^2.1.1",
     "@expo/metro-config": "0.3.16",
-    "@material-ui/core": "^5.0.0-beta.5",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/styles": "^4.11.4",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/lab": "^5.0.0-alpha.50",
-    "@mui/material": "^5.0.2",
-    "@reduxjs/toolkit": "1.8.1",
-    "@testing-library/user-event": "^14.1.0",
-    "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.12",
-    "@types/jsonwebtoken": "^8.5.5",
-    "@types/mongoose": "^5.11.97",
-    "@types/morgan": "^1.9.3",
-    "@ui-kitten/components": "^5.1.1",
-    "@ui-kitten/eva-icons": "^5.1.1",
-    "axios": "^0.21.4",
-    "body-parser": "^1.19.0",
-    "cookie-parser": "^1.4.5",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/lab": "^5.0.0-alpha.81",
+    "@mui/material": "^5.7.0",
+    "@reduxjs/toolkit": "^1.8.1",
+    "@testing-library/user-event": "^14.1.1",
+    "@typescript-eslint/eslint-plugin": "^5.23.0",
+    "@typescript-eslint/parser": "^5.23.0",
+    "@ui-kitten/components": "^5.1.2",
+    "@ui-kitten/eva-icons": "^5.1.2",
+    "cookie-parser": "^1.4.6",
     "core-js": "^3.6.5",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.0",
+    "eslint": "^8.15.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-editorconfig": "^3.2.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "expo": "44.0.6",
     "expo-barcode-scanner": "~11.2.0",
     "expo-dev-client": "0.8.5",
