@@ -1,13 +1,12 @@
 import { StrictMode } from "react";
 
-import { StyledEngineProvider } from "@material-ui/core/styles";
+import { StyledEngineProvider } from "@mui/material/styles";
 import i18n from "i18next";
 import LanguageDetector from "i18next-browser-languagedetector";
 import { initReactI18next } from "react-i18next";
 import { Provider } from "react-redux";
 import { BrowserRouter as Router } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
 import { ThemeProvider } from "styled-components";
 
 import { store } from "../../app/store";
