@@ -1,9 +1,4 @@
 module.exports = {
   stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/preset-create-react-app",
-  ],
-  typescript: { reactDocgen: false },
+  addons: ["@storybook/addon-links", "@storybook/addon-essentials"],
 };
